+++ client/src/app/Paths.ts
@@ -1,14 +1,3 @@
-export const formatPath = (path: Paths, data: any) => {
-  let url = path as string;
-
-  for (const k of Object.keys(data)) {
-    const regex = new RegExp(`:${k}(/|$)`, "g");
-    url = url.replace(regex, data[k] + "$1");
-  }
-
-  return url;
-};
-
 export enum Paths {
   base = "/",
   notFound = "/not-found",
